feat(effects): implement effect system core and editor tooling
- Added EffectFactory with reflection-based registration for Unit and Building effects
- Introduced EffectTypeAttribute to enforce unique mapping between enums and effect classes
- Created EffectData structure for serializable effect configuration
- Implemented EffectDataPropertyDrawer for improved inspector UI
- Added EffectValidator (Editor-only) to ensure enum-to-class consistency
- Defined enums: EffectTarget, UnitEffectType, BuildingEffectType, EffectAreaType
- Added support for assigning multiple effects to each card
- Refactored and organized core effect/data structures into dedicated folders and files
- Implemented basic UI preview widget to display card information
- Renamed EffectAreaParser → EffectAreaCalculator for clarity

+++ Assets/Scripts/Cards/CardInstance.cs
@@ -32,7 +32,7 @@ public abstract class CardInstance
 public class UnitCardInstance : CardInstance
 {
     public int CurrentUnitCount { get; set; }
-    public int CurrentStrength { get; set; }
+    public float CurrentStrength { get; set; }
 
     public UnitCardInstance(UnitCard data, Vector2Int indexCoords) : base(data, indexCoords)
     {
